\name{synRNASeqNet-internal}
\alias{beta_k}
\alias{diagId}
\alias{getIndex}
\alias{getTask}
\alias{shrinkageIntensity}
\alias{startId}
\alias{thetaBayes}
\alias{thetaGT}
\alias{thetaML}
\alias{thetaShrink}
\title{Internal synRNASeqNet Functions}
\description{
 Internal synRNASeqNet Functions
}
\details{
  These are not to be called by the user.
}
\keyword{internal}